MDD/app/requirements.txt
transformers
torch
datasets 
Pillow
matplotlib
Flask
flask-restful
gunicorn